+++ class-ga-measurement-protocol.php
@@ -256,6 +256,7 @@ class Pilau_GA_Measurement_Protocol {
 			'insert-tracking-code'		=> '',
 			'exclude-user-capability'	=> '',
 			'enhanced-link-attribution'	=> '',
+			'allow-anchor-parameters'	=> '',
 			'track-downloads'			=> 'pdf,doc,docx,zip',
 		);
 		$settings = array_merge( $defaults, $settings );
@@ -289,6 +290,7 @@ class Pilau_GA_Measurement_Protocol {
 			$settings['exclude-user-capability'] = preg_replace( '/[^a-z_]/', '', $_POST['exclude-user-capability'] );
 			$settings['track-downloads'] = preg_replace( '/[^a-zA-Z0-9,]/', '', $_POST['track-downloads'] );
 			$settings['enhanced-link-attribution'] = isset( $_POST['enhanced-link-attribution'] ) ? 1 : 0;
+			$settings['allow-anchor-parameters'] = isset( $_POST['allow-anchor-parameters'] ) ? 1 : 0;
 
 			// Save as option
 			$this->set_settings( $settings );
@@ -337,6 +339,11 @@ class Pilau_GA_Measurement_Protocol {
 				$domain = substr( $domain, 4 );
 			}
 
+			// Fields object
+			$fieldsObject = array(
+				'allowAnchor'		=> $this->settings['enhanced-link-attribution']
+			);
+
 			?>
 
 			<!-- Google Analytics tracking inserted by Measurement Protocol plugin -->
@@ -345,7 +352,10 @@ class Pilau_GA_Measurement_Protocol {
 					(i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
 					m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
 				})(window,document,'script','//www.google-analytics.com/analytics.js','ga');
-				ga( 'create', '<?php echo $this->settings['ga-id']; ?>', '<?php echo $domain; ?>' );
+				ga( 'create', '<?php echo $this->settings['ga-id']; ?>', '<?php echo $domain; ?>', { <?php
+					foreach ( $fieldsObject as $fo_key => $fo_value ) { ?>'<?php echo $fo_key ?>': <?php echo $fo_value ? 'true' : 'false'; ?>,
+					<?php } ?>
+				} );
 				<?php if ( $this->settings['enhanced-link-attribution'] ) { ?>
 					ga( 'require', 'linkid', 'linkid.js' );
 				<?php } ?>