+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 0.1.2 (????-??-??)
 * Changed admin headings to H1
+* Added option to `allowAnchor` @link http://z9.io/2010/02/22/remove-unused-utmsource-urls/#comment-672813 @link https://developers.google.com/analytics/devguides/collection/analyticsjs/field-reference#allowAnchor
 
 ## 0.1.1
 * Added POT file for translations